Alphabetize purchasable items
With so many items available for purchase, it can be tricky to find the
right one. Some make a semblance of sense being side-by-side (e.g.
Camelot & Intervale exclusives), but still - given the raw number of
items in this list, alphabetizing makes sense (especially since we
*mostly* had things in order anyway).

In the future, we might just consider making sections to put each item
in.