Remove some unused trip and event fees
The trip fee list is pretty long, it should be helpful to remove items
which are not available for purchase nor likely will be again soon.

Things like the climbers' retreat and the women's open cabin were very
specific one-off gatherings which may have a different fee in the future
(should similar events be held). Other removals, like the No Man's Land
festival, were also one-time events that have long since passed.

We could arguably remove more of these, since prices are subject to
change and it's also unknown when we'll run the next version, but this
is a start.

Most remaining items aren't currently happening, but likely will and are
kept up as examples of what the club does, and likely fees for when
those activities return (e.g. roll clinic & WS).