add intervale covid updates (#57)
* clarify waiver requirements for intervale guests

The Office of General Counsel has requested that we make it explicit
that all guests at Intervale must have signed MITOC's liability waiver.
Thus, we make that more clear in the documentation.

* reduce intervale yurt capacity to 5

The Office of General Counsel has requested that we reduce the limit of
people in the yurt from 10 to 5 in order to allow people more space to
socially distance during covid-19.